fix: change create and delete notice controller path
