+++ src/main/java/com/ajou/travely/controller/notice/NoticeController.java
@@ -25,12 +25,6 @@ import java.util.Optional;
 public class NoticeController {
     private final NoticeService noticeService;
 
-    @PostMapping(value = "", consumes = "multipart/form-data")
-    public ResponseEntity<NoticeResponseDto> createNotice(@LoginUser SessionUser sessionUser,
-                                                          @Valid @ModelAttribute NoticeCreateRequestDto requestDto) {
-        return ResponseEntity.ok(noticeService.createNotice(sessionUser.getUserId(), requestDto));
-    }
-
     @GetMapping("/{noticeId}")
     public ResponseEntity<NoticeResponseDto> getNotice(@PathVariable Long noticeId) {
         return ResponseEntity.ok(noticeService.getNotice(noticeId));
@@ -46,9 +40,5 @@ public class NoticeController {
         return ResponseEntity.ok(notices);
     }
 
-    @DeleteMapping("/{noticeId}")
-    public ResponseEntity<Void> deleteNotice(@PathVariable Long noticeId) {
-        noticeService.deleteNotice(noticeId);
-        return ResponseEntity.ok().build();
-    }
+
 }
